Organize user DTOs into user package
