+++ modules/fbs-identity-service/api/src/main/kotlin/de/thm/ii/fbs/fbs_identity_service/dto/login/LoginRequest.kt
@@ -1,0 +1,6 @@
+package de.thm.ii.fbs.fbs_identity_service.dto.login
+
+data class LoginRequest(
+    val username: String,
+    val password: String
+)
