+++ modules/fbs-identity-service/api/src/main/kotlin/de/thm/ii/fbs/fbs_identity_service/dto/login/LoginResponse.kt
@@ -1,0 +1,7 @@
+package de.thm.ii.fbs.fbs_identity_service.dto.login
+
+data class LoginResponse(
+    val accessToken: String,
+    val tokenType: String = "Bearer",
+    val expiresIn: Long
+)
