+++ modules/fbs-identity-service/api/src/main/kotlin/de/thm/ii/fbs/fbs_identity_service/dto/user/ChangeUserPasswordInput.kt
@@ -1,4 +1,4 @@
-package de.thm.ii.fbs.fbs_identity_service.dto
+package de.thm.ii.fbs.fbs_identity_service.dto.user
 
 data class ChangeUserPasswordInput(
     val userId: Long,
